UPSTREAM: <carry>: Add OpenShift specific files
Signed-off-by: dtfranz <dfranz@redhat.com>

UPSTREAM: <carry>: Update generate-manifests.sh

Recent upstream changes to catalogd moved the location of the default
kustomization directory.
It used to be `config/default`, and is now `config/base/default`.
The `openshift/generate-manifests.sh` script assumes the original
location, this update will check for the location of the default
directory.

Signed-off-by: Todd Short <todd.short@me.com>